index bug fixed. similar molecule search now supports index. configurable threshold removed. bad idea.

+++ CGRdb/sql/similar_molecule.sql
@@ -17,7 +17,7 @@
 */
 
 CREATE OR REPLACE FUNCTION
-"{schema}".cgrdb_search_similar_molecules(data bytea, threshold float, OUT id integer, OUT count integer)
+"{schema}".cgrdb_search_similar_molecules(data bytea, OUT id integer, OUT count integer)
 AS $$
 from CGRtools.containers import MoleculeContainer
 from compress_pickle import loads
@@ -40,19 +40,37 @@ if found:
 # cache not found. lets start searching
 fp = GD['cgrdb_mfp'].transform_bitset([molecule])[0]
 
-plpy.execute('DROP TABLE IF EXISTS cgrdb_query')
-plpy.execute(f'''CREATE TEMPORARY TABLE cgrdb_query ON COMMIT DROP AS
+if GD['index']:  # use index search
+    from requests import post
+    found = post(f"{GD['index']}/similarity/molecule", json=fp).json()
+    if found:  # create cgrdb_query temp table
+        plpy.execute('DROP TABLE IF EXISTS cgrdb_query')
+        if isinstance(found[0], int):  # need to calculate tanimoto
+            plpy.execute(f'''CREATE TEMPORARY TABLE cgrdb_query ON COMMIT DROP AS
+SELECT x.molecule m,
+       icount(x.fingerprint & ARRAY{fp}::integer[])::float / icount(x.fingerprint | ARRAY{fp}::integer[])::float t
+FROM "{schema}"."MoleculeStructure" x
+WHERE x.id = ANY(ARRAY{found}::integer[])''')
+        else:  # tanimoto exists
+            plpy.execute(f'''CREATE TEMPORARY TABLE cgrdb_query ON COMMIT DROP AS
+SELECT x.molecule m, f.t
+FROM "{schema}"."MoleculeStructure" x,
+     (VALUES {', '.join(f'({s}::integer, {t:.2f}::float)' for s, t in found)}) AS f (s, t)
+WHERE x.id = f.s''')
+else:  # sequential search
+    plpy.execute('DROP TABLE IF EXISTS cgrdb_query')
+    plpy.execute(f'''CREATE TEMPORARY TABLE cgrdb_query ON COMMIT DROP AS
 SELECT c.m, c.t
 FROM (
     SELECT x.molecule m,
            icount(x.fingerprint & ARRAY{fp}::integer[])::float / icount(x.fingerprint | ARRAY{fp}::integer[])::float t
     FROM "{schema}"."MoleculeStructure" x
-    WHERE x.fingerprint && ARRAY{fp}::integer[]
 ) c
-WHERE c.t > {threshold}''')
+WHERE c.t > 0.5''')
+    # check for empty results
+    found = plpy.execute('SELECT COUNT(*) FROM cgrdb_query')[0]['count']
 
-# check for empty results
-if not plpy.execute('SELECT COUNT(*) FROM cgrdb_query')[0]['count']:
+if not found:
     # store empty cache
     found = plpy.execute(f'''INSERT INTO
 "{schema}"."MoleculeSearchCache"(signature, operator, date, molecules, tanimotos)
